tedla-hypertension-project-databricks/src/nlp_method/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
blis==0.7.11
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm==0.2.3
confection==0.1.5
cymem==2.0.13
Cython==3.1.7
dask==2025.11.0
dataframe_api_compat==0.2.7
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distributed==2025.11.0
duckdb==1.4.2
en_core_web_sm @ file:///home/westerd/_/research_projects/tedla-hypertension/.models/en_core_web_sm-3.7.1.tar.gz#sha256=1075c2aa2bc2fee105ab6e90a01a5d1a428c9f5b20a1fa003dc2cb6a438d295e
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
fqdn==1.5.1
fsspec==2025.12.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
invoke==2.2.1
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
langcodes==3.5.0
language_data==1.4.0
lark==1.3.1
locket==1.0.0
loguru==0.7.3
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
medspacy==1.3.1
medspacy_quickumls==3.2
medspacy_unqlite==0.9.8
mistune==3.1.4
modin==0.37.1
modin-spreadsheet==0.1.2
msgpack==1.1.2
murmurhash==1.0.15
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.2
notebook==7.5.0
notebook_shim==0.2.4
numpy==1.26.4
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
partd==1.4.2
pexpect==4.9.0
platformdirs==4.5.0
pluggy==1.6.0
preshed==3.0.12
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyFastNER==1.0.10
Pygments==2.19.2
pyodbc==5.3.0
PyRuSH==1.0.12
pysbd==0.3.4
pysimstring==1.3.0
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
quicksectx==0.4.1
ray==2.52.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.30.0
Send2Trash==1.8.3
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sortedcontainers==2.4.0
soupsieve==2.8
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
srsly==2.5.2
stack-data==0.6.3
tblib==3.2.2
terminado==0.18.1
thinc==8.2.5
tinycss2==1.4.0
toolz==1.1.0
tornado==6.5.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
traitlets==5.14.3
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
uri-template==1.3.0
urllib3==2.5.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.3
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.15
wrapt==2.0.1
zict==3.0.0
zipp==3.23.0
